refactor: [FE] 에디터 이벤트 핸들러 분리 에디터 이벤트 핸들러 Editor 컴포넌트 하위로 이동
